Moving the module to its own file
